chore(lint): move Cypress namespace to index.d.ts
… to please linter ("no-namespace" rule)
See: https://github.com/cypress-io/cypress-documentation/issues/4043#issuecomment-1046404794
